+++ src/Pinch.hs
@@ -148,9 +148,11 @@ module Pinch
     ) where
 
 import Control.Monad
-import Data.ByteString (ByteString)
-import Data.Int        (Int32)
-import Data.Text       (Text)
+import Data.ByteString    (ByteString)
+import Data.Int           (Int32)
+import Data.Serialize.Get (runGetState)
+import Data.Text          (Text)
+import Data.Tuple         (swap)
 
 import Pinch.Internal.Builder   (runBuilder)
 import Pinch.Internal.Generic
@@ -217,7 +219,7 @@ decode p = deserializeValue p >=> runParser . unpinch
 -- Right ("\NUL\NUL\NUL",[1,2,3,4,5])
 --
 decodeWithLeftovers :: Pinchable a => Protocol -> ByteString -> Either String (ByteString, a)
-decodeWithLeftovers p = deserializeValue' p >=> traverse (runParser . unpinch)
+decodeWithLeftovers p bs = (swap <$> runGetState (deserializeValue' p) bs 0) >>= traverse (runParser . unpinch)
 {-# INLINE decodeWithLeftovers #-}
 
 ------------------------------------------------------------------------------